backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.9.0
